SFML fully functional
There are some big changes in this commit. First of all, the SFML library and its audio, graphics, window, system submodules have been linked successfully. A new audio namespace is introduced for audio development and its classes. AudioHandler class was created. As the name suggests, it handles the audio -  it can load and unload audio and music files, throw exceptions on demand and create threads for faster audio utilization.