Successfully implemented Audio Playback
The audio playback of .wav files works properly. However, there are multiple problems: The main thread stops responding after the sound starts playing, the default audio device is selected on Linux, which might not be the device you are currently using. Next step is fixing those problems and adding more abstraction to the class for stopping, seeking and changing the audio.